+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+
+* Fix crash when `ThreeDSecureRequest` `amount` field is set to NaN (resolves #507)
+
 ## 4.32.0 (2020-02-18)
 
 * Update CardinalMobile.framework to v2.2.2-1